(Main v2.0) Reimplemented Ability System
- Reimplemented ability system properly this time.
- Everything is working perfectly fine here.

+++ Assets/Scripts/GameScripts/Interfaces/IDamageable.cs
@@ -1,7 +1,7 @@
 public interface IDamageable
 {
-    void TakeDamage(int damageValue);
-    int GetCurrentHealth();
+    void TakeDamage(float damageValue);
+    float GetCurrentHealth();
     bool IsAlive();
 
 }
